yarn+project structure+react
- added yarn
- adding project structure
- added React (may remove later if not used)
- refactored sop
