+++ ReactView.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export const ReactView = () => {
+	return <h4> Hello, React!</h4>
+}
