+++ manifest.json
@@ -1,11 +1,9 @@
 {
-	"id": "obsidian-sample-plugin",
-	"name": "Sample Plugin",
-	"version": "1.0.0",
+	"id": "cabin-flow-plugin",
+	"name": "Cabin Flow Plugin",
+	"version": "0.1.0",
 	"minAppVersion": "0.15.0",
-	"description": "This is a sample plugin for Obsidian. This plugin demonstrates some of the capabilities of the Obsidian API.",
-	"author": "Obsidian",
-	"authorUrl": "https://obsidian.md",
-	"fundingUrl": "https://obsidian.md/pricing",
+	"description": "Trello <> Obsidian <> Todoist",
+	"author": "Raghu Mulukutla",
 	"isDesktopOnly": false
 }
