+++ package.json
@@ -22,6 +22,10 @@
 		"typescript": "4.7.4"
 	},
 	"dependencies": {
-		"eslint": "^8.36.0"
+		"@types/react": "^18.0.29",
+		"@types/react-dom": "^18.0.11",
+		"eslint": "^8.36.0",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0"
 	}
 }
